Move remaining display functions out of panels.cpp
